pwcbackend-main/requirements.txt
Django>=6.0,<7.0
django-cors-headers>=4.9.0,<5.0.0
djangorestframework>=3.17.1,<4.0.0
psycopg>=3.3.3,<4.0.0
